+++ macros/src/scheme_derive/gen_config_struct.rs
@@ -10,6 +10,7 @@ pub fn generate_config_struct(parsed: &ParsedScheme) -> syn::Result<TokenStream>
         config_struct_name,
         basis,
         commands,
+        config_ext,
         ..
     } = parsed;
     let command_names_snake = commands
@@ -17,6 +18,7 @@ pub fn generate_config_struct(parsed: &ParsedScheme) -> syn::Result<TokenStream>
         .map(|c| &c.command_name_snake)
         .collect::<Vec<_>>();
     let action_types = commands.iter().map(|c| c.action_type).collect::<Vec<_>>();
+    let config_ext = config_ext.iter();
     Ok(quote! {
         #[derive(bevy::prelude::Asset, bevy::prelude::TypePath, bevy_tnua::serde::Serialize, bevy_tnua::serde::Deserialize)]
         #[serde(crate = "bevy_tnua::serde")]
@@ -24,6 +26,9 @@ pub fn generate_config_struct(parsed: &ParsedScheme) -> syn::Result<TokenStream>
             #vis basis: <#basis as bevy_tnua::TnuaBasis>::Config,
             #(
                 #vis #command_names_snake: <#action_types as bevy_tnua::TnuaAction<#basis>>::Config,
+            )*
+            #(
+                #vis ext: #config_ext,
             )*
         }